Move meltdown check to base system
